chore: working on parsing

+++ src/types.rs
@@ -3,4 +3,5 @@ pub(crate) static MOD_STATE_FIELD: u32 = 255;
 pub(crate) static ESCAPE_KEYSYM: u32 = 0xff1b;
 pub(crate) static SYNCHRONOUS_CHAR: char = ';';
 
-pub(crate) type Xid = i32;
+// pub(crate) type Xid = i32;
+pub(crate) type Xid = u8;
